workspace/tooth-ai/requirements_full.txt
# Tooth-AI Full Requirements
# Core dependencies for complete system

# Numerical computing
numpy>=1.24.0
scipy>=1.10.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Model libraries
timm>=0.9.0

# Detectron2 (install separately if needed)
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# UI framework
streamlit>=1.28.0

# COCO tools
pycocotools>=2.0.6

# ONNX
onnx>=1.14.0
onnxruntime>=1.15.0

# TensorRT (optional, may require separate installation)
# tensorrt>=8.6.0
# pycuda>=2022.2.2

# System utilities
psutil>=5.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
pandas>=2.0.0
scikit-learn>=1.3.0

# YAML parsing
pyyaml>=6.0

# Utilities
tqdm>=4.66.0



